p3migration: 'float' object cannot be interpreted as an integer

+++ src/busses/wishbone/wishbone.py
@@ -181,7 +181,7 @@ def genCaseByteEnable(masterinterface):
         if int(size) == int(master_size):
             out += write_bus_name
         else:
-            bitsize = byte_en_size / nb_byte
+            bitsize = int(byte_en_size / nb_byte)
             for i in range(bitsize):
                 val = bin(int(mask * pow(2, nb_byte*i)))[2:]
                 out += write_bus_name + "(" + \
@@ -206,10 +206,10 @@ def genCaseByteEnable(masterinterface):
             if not (int(size) == int(master_size)):
                 nb_byte = size / 8
                 mask = pow(2, nb_byte)-1
-                for i in range(byte_en_size / nb_byte):
+                for i in range(int(byte_en_size / nb_byte)):
                     val = bin(int(mask * pow(2, nb_byte*i)))[2:]
                     out += ' "' + \
-                        str(((bin(i*(size/8))[2:]).zfill(3))) + \
+                        str(((bin(i*(int(size / 8)))[2:]).zfill(3))) + \
                         '" when ' + byte_en_name + ' = "' + \
                         val.zfill(byte_en_size) + '" else\n' + 2 * ONETAB
 
@@ -557,7 +557,7 @@ def selectWrite(masterinterface, intercon):
         else:
             nb_byte = data_size / 8
             mask = pow(2, nb_byte)-1
-            bitsize = master_size / data_size
+            bitsize = int(master_size / data_size)
             for i in range(bitsize):
                 min_pos = i * data_size
                 val = bin(int(mask * pow(2, nb_byte*i)))[2:]
